Conflicts:
	GDALmake.opt.in
	NEWS
	VERSION
	alg/gdalgrid_priv.h
	alg/gdaltransformer.cpp
	alg/llrasterize.cpp
	apps/gdalbuildvrt.cpp
	apps/gdaldem.cpp
	apps/gdalserver.c
	apps/gdalwarp.cpp
	apps/ogr2ogr.cpp
	configure.in
	data/ruian_vf_ob_v1.gfs
	data/ruian_vf_v1.gfs
	doxygen_sqlite3.db
	frmts/bsb/bsb_read.c
	frmts/envisat/EnvisatFile.c
	frmts/ers/ersdataset.cpp
	frmts/frmt_various.html
	frmts/grass/pkg/configure.in
	frmts/gtiff/geotiff.cpp
	frmts/gtiff/gt_wkt_srs.cpp
	frmts/hdf4/hdf4dataset.cpp
	frmts/hf2/hf2dataset.cpp
	frmts/hfa/hfaband.cpp
	frmts/hfa/hfaentry.cpp
	frmts/netcdf/netcdfdataset.cpp
	frmts/nitf/ecrgtocdataset.cpp
	frmts/nitf/nitffile.c
	frmts/nitf/rpftocfile.cpp
	frmts/northwood/grcdataset.cpp
	frmts/northwood/grddataset.cpp
	frmts/northwood/northwood.cpp
	frmts/pdf/pdfdataset.cpp
	frmts/pdf/pdfio.cpp
	frmts/pdf/pdfio.h
	frmts/pds/pdsdataset.cpp
	frmts/raw/btdataset.cpp
	frmts/til/tildataset.cpp
	frmts/vrt/vrtdataset.cpp
	frmts/vrt/vrtrasterband.cpp
	frmts/vrt/vrtrawrasterband.cpp
	frmts/vrt/vrtsources.cpp
	frmts/wms/wmsdriver.cpp
	frmts/xpm/xpmdataset.cpp
	gcore/gdal_version.h
	gcore/gdaljp2metadata.cpp
	gcore/gdalproxypool.cpp
	gcore/gdalrasterband.cpp
	man/man1/_home_rouault_dist_wrk_gdal_apps_.1
	man/man1/gdal-config.1
	man/man1/gdal2tiles.1
	man/man1/gdal_calc.1
	man/man1/gdal_contour.1
	man/man1/gdal_edit.1
	man/man1/gdal_fillnodata.1
	man/man1/gdal_grid.1
	man/man1/gdal_merge.1
	man/man1/gdal_polygonize.1
	man/man1/gdal_proximity.1
	man/man1/gdal_rasterize.1
	man/man1/gdal_retile.1
	man/man1/gdal_sieve.1
	man/man1/gdal_translate.1
	man/man1/gdal_utilities.1
	man/man1/gdaladdo.1
	man/man1/gdalbuildvrt.1
	man/man1/gdalcompare.1
	man/man1/gdaldem.1
	man/man1/gdalinfo.1
	man/man1/gdallocationinfo.1
	man/man1/gdalmanage.1
	man/man1/gdalmove.1
	man/man1/gdalsrsinfo.1
	man/man1/gdaltindex.1
	man/man1/gdaltransform.1
	man/man1/gdalwarp.1
	man/man1/nearblack.1
	man/man1/ogr2ogr.1
	man/man1/ogr_utilities.1
	man/man1/ogrinfo.1
	man/man1/ogrlineref.1
	man/man1/ogrtindex.1
	man/man1/pct2rgb.1
	man/man1/rgb2pct.1
	ogr/ogr_srs_api.h
	ogr/ogr_srs_esri.cpp
	ogr/ogrgeometrycollection.cpp
	ogr/ogrsf_frmts/csv/ogrcsvlayer.cpp
	ogr/ogrsf_frmts/edigeo/ogredigeodatasource.cpp
	ogr/ogrsf_frmts/generic/ogr_gensql.cpp
	ogr/ogrsf_frmts/generic/ogrlayer.cpp
	ogr/ogrsf_frmts/geojson/ogrgeojsondatasource.cpp
	ogr/ogrsf_frmts/geojson/ogrgeojsonreader.cpp
	ogr/ogrsf_frmts/geojson/ogrtopojsonreader.cpp
	ogr/ogrsf_frmts/gml/gmlhandler.cpp
	ogr/ogrsf_frmts/gml/parsexsd.cpp
	ogr/ogrsf_frmts/gpkg/ogrgeopackagelayer.cpp
	ogr/ogrsf_frmts/gpkg/ogrgeopackageutility.cpp
	ogr/ogrsf_frmts/mitab/mitab_feature_mif.cpp
	ogr/ogrsf_frmts/mssqlspatial/ogrmssqlspatialtablelayer.cpp
	ogr/ogrsf_frmts/ntf/ntfrecord.cpp
	ogr/ogrsf_frmts/openfilegdb/filegdbtable.cpp
	ogr/ogrsf_frmts/openfilegdb/ogropenfilegdbdatasource.cpp
	ogr/ogrsf_frmts/openfilegdb/ogropenfilegdblayer.cpp
	ogr/ogrsf_frmts/pg/ogrpgtablelayer.cpp
	ogr/ogrsf_frmts/pgdump/ogrpgdumpdatasource.cpp
	ogr/ogrsf_frmts/pgdump/ogrpgdumplayer.cpp
	ogr/ogrsf_frmts/shape/shape2ogr.cpp
	ogr/ogrsf_frmts/sqlite/ogrsqliteselectlayer.cpp
	ogr/ogrsf_frmts/vrt/ogrvrtlayer.cpp
	port/cpl_path.cpp
	port/cpl_string.cpp
	port/cpl_vsil_stdout.cpp
	port/cpl_vsil_tar.cpp
	swig/include/perl/gdal_perl.i
	swig/include/perl/ogr_perl.i
	swig/include/python/ogr_python.i
	swig/perl/gdalconst_wrap.c
	swig/perl/lib/Geo/GDAL.pm
	swig/perl/lib/Geo/OGR.dox
	swig/perl/lib/Geo/OGR.pm
	swig/perl/ogr_wrap.cpp
	swig/perl/osr_wrap.cpp
	swig/perl/t/gdal.t
	swig/perl/t/ogr.t
	swig/python/samples/ogr_layer_algebra.py
	swig/python/scripts/gdal_merge.py
	swig/python/setup.py

+++ swig/java/apps/ogrtindex.java
@@ -1,5 +1,5 @@
 /******************************************************************************
- * $Id: ogrtindex.java 23738 2012-01-09 19:40:54Z rouault $
+ * $Id: ogrtindex.java 28375 2015-01-30 12:06:11Z rouault $
  *
  * Project:  OpenGIS Simple Features Reference Implementation
  * Purpose:  Program to generate a UMN MapServer compatible tile index for a
@@ -250,7 +250,7 @@ public class ogrtindex {
       String[] existingLayersTab = null;
       SpatialReference alreadyExistingSpatialRef = null;
       boolean alreadyExistingSpatialRefValid = false;
-      nExistingLayers = poDstLayer.GetFeatureCount();
+      nExistingLayers = (int)poDstLayer.GetFeatureCount();
       if (nExistingLayers > 0)
       {
          existingLayersTab = new String[nExistingLayers];
